fix: theme add submit 오류 픽스
