+++ app/(shared)/(ThemeTextField)/Container.tsx
@@ -23,15 +23,10 @@ export default function ThemeTextField({
     inputRef,
     handleInputChange,
     handleInputBlur,
-    handleKeyDownSubmit,
   } = useTextField({ id, content, checkErrorText });
 
   return (
-    <div
-      tabIndex={isFocus ? -1 : tabIndex}
-      onFocus={() => setIsFocus(true)}
-      onKeyDown={handleKeyDownSubmit}
-    >
+    <div tabIndex={isFocus ? -1 : tabIndex} onFocus={() => setIsFocus(true)}>
       <button
         className={classNames("theme-textfield", {
           focus: isFocus,
