+++ app/admin-new/(components)/CreateTheme/CreateThemeBody.tsx
@@ -1,5 +1,4 @@
 import React from "react";
-import { useCreateThemeReset } from "@/components/atoms/createTheme.atom";
 import ThemeTextField from "../../../(shared)/(ThemeTextField)/Container";
 import {
   hintCountTextFieldProps,
@@ -8,9 +7,6 @@ import {
 } from "./createTheme";
 
 export default function CreateThemeBody() {
-  const resetCreateTheme = useCreateThemeReset();
-
-  resetCreateTheme();
   return (
     <div className="create-theme__body">
       <ThemeTextField {...nameTextFieldProps} />
